feature/media-query -> Updating media query
Checking if there is a media in the list before returning it and returning the DTO related to the engine (correcting mappign problems between MediaPublicDto and MediaUnityModel for example). Also updating return type of media query route because it now returns any type of engine media. There should be a better type, checking it later